# src/app/static/Swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Documentación de la API",
    "description": "API para gestionar roles y usuarios, incluyendo autenticación y notificación por correo electrónico",
    "version": "1.0"
  },
  "host": "localhost:4000",
  "schemes": ["http"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
    }
  },
  "paths": {
    "/register": {
      "post": {
        "summary": "Registrar un nuevo usuario",
        "description": "Crea una cuenta de usuario nueva. No requiere autenticación.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": ["email", "password"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usuario registrado con éxito"
          },
          "409": {
            "description": "El correo ya existe"
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Iniciar sesión",
        "description": "Autentica a un usuario y devuelve un token JWT. No requiere autenticación.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": ["email", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Autenticación exitosa, token proporcionado"
          },
          "401": {
            "description": "Credenciales inválidas"
          }
        }
      }
    },
...
  }
}